extract sound manager
